+++ changelog.md
@@ -1,11 +1,12 @@
 # Changelog
 
 ### Version 0.2.0 (Beta-3) - TBA
-- New Feature: Introduced paging
+- New Feature: Introduced paging (results > memory could be requested)
 - New Frature: Introduced type safe client API
 - Improvement: Simplified network protocol by removing the duplicate success and response packages
 - Improvement: Success and Error responses are now indicated by the isFailed() future method 
 - Improvement: Success and Error messages are now provided by the getMessage() future method 
+- Improvement: Introduced iterator for list results in client API
 - Bugfix: Group membership watch was not re-established
 - Bugfix: Wrong node state was displayed in the GUI
 - Bugfix: Fixed handling of failed futures in the benchmark